+++ docs/content/en/changelog.md
@@ -14,8 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 **Added**
 
-- `toBeSameModelAs()` expectation
 - `toBeBirthday()` expectation
+- `toBeSameModelAs()` expectation
+- `toBeMidday()` expectation
 
 ### [v1.5.0](https://github.com/pestphp/defstudio-plugin-laravel-expectations/compare/v1.4.0...v1.5.0) - 2021-10-25